extract typings into a different file
